+++ context/app/static/js/components/detailPage/provenance/ProvTableDerivedLink/ProvTableDerivedLink.jsx
@@ -1,9 +1,18 @@
 import React from 'react';
 
+import { useTrackEntityPageEvent } from 'js/components/detailPage/useTrackEntityPageEvent';
 import { LinkButton } from './style';
 
 function ProvTableDerivedLink({ uuid, type }) {
-  return <LinkButton href={`/search?ancestor_ids[0]=${uuid}&entity_type[0]=${type}`}>View Derived {type}s</LinkButton>;
+  const trackEntityPageEvent = useTrackEntityPageEvent();
+  return (
+    <LinkButton
+      href={`/search?ancestor_ids[0]=${uuid}&entity_type[0]=${type}`}
+      onClick={() => trackEntityPageEvent({ action: 'Provenance / Table / View Derived' })}
+    >
+      View Derived {type}s
+    </LinkButton>
+  );
 }
 
 export default ProvTableDerivedLink;