+++ context/app/static/js/components/detailPage/summary/SummaryJSONButton/index.ts
@@ -1,0 +1,3 @@
+import SummaryJSONButton from './SummaryJSONButton';
+
+export default SummaryJSONButton;