+++ context/app/static/js/pages/Sample/Sample.jsx
@@ -11,7 +11,6 @@ import SummaryItem from 'js/components/detailPage/summary/SummaryItem';
 import DetailLayout from 'js/components/detailPage/DetailLayout';
 import MetadataTable from 'js/components/detailPage/MetadataTable';
 import SampleTissue from 'js/components/detailPage/SampleTissue';
-import useSendUUIDEvent from 'js/components/detailPage/useSendUUIDEvent';
 import useEntityStore from 'js/stores/useEntityStore';
 import { DetailContext } from 'js/components/detailPage/DetailContext';
 import { getSectionOrder } from 'js/components/detailPage/utils';
@@ -19,6 +18,7 @@ import { getSectionOrder } from 'js/components/detailPage/utils';
 import DerivedDatasetsSection from 'js/components/detailPage/derivedEntities/DerivedDatasetsSection';
 
 import { combineMetadata } from 'js/pages/utils/entity-utils';
+import useTrackID from 'js/hooks/useTrackID';
 
 const entityStoreSelector = (state) => state.setAssayMetadata;
 
@@ -60,7 +60,7 @@ function SampleDetail() {
     setAssayMetadata({ hubmap_id, entity_type, mapped_organ, sample_category });
   }, [hubmap_id, entity_type, mapped_organ, sample_category, setAssayMetadata]);
 
-  useSendUUIDEvent(entity_type, uuid);
+  useTrackID({ entity_type, hubmap_id });
 
   const detailContext = useMemo(() => ({ hubmap_id, uuid }), [hubmap_id, uuid]);